Sorting Update
List can now be sorted to total playtime and then back to alphabetical (which is what it'll be sorted by when opening the page the first time)

Somethings have been moved to different files, but the code itself hasn't changed much.

There is now a <script> tag at the bottom of the page that hold the Javascript/JQuery that now builds the list dynamically, which is why it can now be updated!
The <script> tag can be found in the foot.php file